@@ -3,46 +3,46 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.6.4",
-    "@hookform/resolvers": "^2.8.3",
+    "@ant-design/icons": "^4.7.0",
+    "@hookform/resolvers": "^2.8.8",
     "@material-ui/icons": "^4.11.2",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/howler": "^2.2.4",
-    "@types/jest": "^26.0.24",
-    "@types/material-ui": "^0.21.9",
-    "@types/node": "^12.20.23",
+    "@reduxjs/toolkit": "^1.8.0",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/howler": "^2.2.7",
+    "@types/jest": "^27.4.1",
+    "@types/material-ui": "^0.21.12",
+    "@types/node": "^17.0.23",
     "@types/offscreencanvas": "^2019.6.4",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
-    "@types/react-router-dom": "^5.1.8",
-    "antd": "^4.16.13",
-    "axios": "^0.21.4",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
+    "@types/react-redux": "^7.1.23",
+    "@types/react-router-dom": "^5.3.3",
+    "antd": "^4.19.3",
+    "axios": "^0.26.1",
     "classnames": "^2.3.1",
     "easy-yandex-s3": "^1.1.6",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-prettier": "^8.5.0",
     "howler": "^2.2.3",
     "node-sass": "^7.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.18.1",
+    "react-hook-form": "^7.28.1",
     "react-icons": "^4.3.1",
     "react-image-crop": "^9.0.5",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3",
-    "redux": "^4.1.1",
-    "redux-thunk": "^2.3.0",
-    "typescript": "^4.4.2",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.2",
+    "react-scripts": "^5.0.0",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
+    "typescript": "^4.6.3",
     "uuid": "^8.3.2",
-    "web-vitals": "^1.1.2",
+    "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
+    "start": "react-scripts start",
     "build": "CI=false && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -66,6 +66,7 @@
     ]
   },
   "devDependencies": {
-    "@types/uuid": "^8.3.3"
+    "@types/uuid": "^8.3.4",
+    "redux-mock-store": "^1.5.4"
   }
 }
