@@ -6,7 +6,9 @@ const instance = axios.create({
 });
 
 instance.interceptors.request.use(function (config: AxiosRequestConfig) {
-    config.headers.Authorization = `Bearer ${localStorage.getItem('token')}`;
+    if (config.headers) {
+        config.headers.Authorization = `Bearer ${localStorage.getItem('token')}`;
+    }
     return config;
 });
 
