@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
 import reportWebVitals from './reportWebVitals';
@@ -11,11 +11,11 @@ import './index.css';
 
 ReactDOM.render(
 	<React.StrictMode>
-		<Router>
-			<Provider store={store}>
+		<Provider store={store}>
+			<BrowserRouter>
 				<App />
-			</Provider>
-		</Router>
+			</BrowserRouter>
+		</Provider>
 	</React.StrictMode>,
 	document.getElementById('root'),
 );
